refactor: SearchEngineImpl 내 OOP 관점에서 분할 정의
