+++ gateway/src/main/kotlin/me/ogq/ocp/sample/gateway/elasticsearch/EsRequest.kt
@@ -1,0 +1,55 @@
+package me.ogq.ocp.sample.gateway.elasticsearch
+
+import me.ogq.ocp.sample.model.publicityright.PublicityRight
+import org.elasticsearch.action.search.SearchRequest
+import org.elasticsearch.index.query.BoolQueryBuilder
+import org.elasticsearch.index.query.Operator
+import org.elasticsearch.index.query.QueryBuilders
+import org.elasticsearch.search.builder.SearchSourceBuilder
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.stereotype.Component
+
+@Component
+class EsRequest(
+    @Value("\${spring.elasticsearch.index}")
+    private val indexName: String,
+) {
+    fun createWith(
+        publicityRight: PublicityRight?,
+        query: String,
+        page: Int,
+        pageSize: Int
+    ): SearchRequest {
+        val queryBuilder = QueryBuilders.boolQuery()
+        addQueryStringTermQuery(query, queryBuilder)
+        addPublicityRightTermQuery(publicityRight, queryBuilder)
+
+        val sourceBuilder = SearchSourceBuilder()
+        sourceBuilder.query(queryBuilder)
+        sourceBuilder.from(page * pageSize)
+        sourceBuilder.size(pageSize)
+
+        return SearchRequest(indexName).source(sourceBuilder)
+    }
+
+    private fun addQueryStringTermQuery(query: String, queryBuilder: BoolQueryBuilder) {
+        queryBuilder.must(
+            QueryBuilders.queryStringQuery("*$query*")
+                .field("title")
+                .field("tags")
+                .lenient(true)
+                .defaultOperator(Operator.OR)
+        )
+    }
+
+    private fun addPublicityRightTermQuery(
+        publicityRight: PublicityRight?,
+        queryBuilder: BoolQueryBuilder
+    ) {
+        if (publicityRight != null) {
+            queryBuilder.should(QueryBuilders.termQuery("publicity_id", publicityRight.id.toString()))
+        } else {
+            queryBuilder.should(QueryBuilders.boolQuery().mustNot(QueryBuilders.existsQuery("publicity_id")))
+        }
+    }
+}
